+++ doczrc.js
@@ -1,5 +1,6 @@
 export default {
   title: 'MDwalks-UI',
+  description: 'MDwalks-UI Component document', 
   native: true,
   typescript: false,
   port: '9004',
@@ -11,5 +12,7 @@ export default {
   },
   modifyBundlerConfig: (config) => {
     return config
-  }
+  },
+  dest: '/docs',
+  base: '/MDwalks-UI/' // gatsby-config.tpl.js 참고
 }